fix(material/schematics): fix icon button selector and move button migrator to its own folder
