+++ src/material/schematics/ng-generate/mdc-migration/rules/index.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {ButtonStylesMigrator} from './button-styles';
+import {ButtonStylesMigrator} from './components/button/button-styles';
 import {StyleMigrator} from './style-migrator';
 
 export const MIGRATORS: StyleMigrator[] = [new ButtonStylesMigrator()];
